workload: remove workload/cluster
Removes workload/cluster and replaces it with workloadv2/cluster, which
uses server.Server.
