+++ tests/proxy_test.go
@@ -16,7 +16,7 @@ import (
 	"testing"
 
 	"github.com/andydunstall/piko/agent/client"
-	"github.com/andydunstall/piko/workload/cluster"
+	cluster "github.com/andydunstall/piko/workloadv2/cluster"
 	"github.com/gorilla/websocket"
 	"github.com/stretchr/testify/assert"
 )
@@ -28,7 +28,7 @@ type errorMessage struct {
 // Tests proxying HTTP traffic with a single Piko server node.
 func TestProxy_HTTP(t *testing.T) {
 	t.Run("http", func(t *testing.T) {
-		node := cluster.NewNodeV2()
+		node := cluster.NewNode()
 		node.Start()
 		defer node.Stop()
 
@@ -78,7 +78,7 @@ func TestProxy_HTTP(t *testing.T) {
 	})
 
 	t.Run("https", func(t *testing.T) {
-		node := cluster.NewNodeV2(cluster.WithTLS(true))
+		node := cluster.NewNode(cluster.WithTLS(true))
 		node.Start()
 		defer node.Stop()
 
@@ -139,7 +139,7 @@ func TestProxy_HTTP(t *testing.T) {
 	})
 
 	t.Run("websocket", func(t *testing.T) {
-		node := cluster.NewNodeV2()
+		node := cluster.NewNode()
 		node.Start()
 		defer node.Stop()
 
@@ -188,7 +188,7 @@ func TestProxy_HTTP(t *testing.T) {
 
 	// Tests sending a request to an endpoint with no listeners.
 	t.Run("no listeners", func(t *testing.T) {
-		node := cluster.NewNodeV2()
+		node := cluster.NewNode()
 		node.Start()
 		defer node.Stop()
 
@@ -214,7 +214,7 @@ func TestProxy_HTTP(t *testing.T) {
 
 func TestProxy_TCP(t *testing.T) {
 	t.Run("tcp", func(t *testing.T) {
-		node := cluster.NewNodeV2()
+		node := cluster.NewNode()
 		node.Start()
 		defer node.Stop()
