+++ workloadv2/cluster/node.go
@@ -2,10 +2,15 @@ package cluster
 
 import (
 	"context"
+	"crypto/rsa"
+	"crypto/x509"
+	"encoding/pem"
+	"os"
 	"sync"
 	"time"
 
 	"github.com/andydunstall/piko/pkg/log"
+	"github.com/andydunstall/piko/pkg/testutil"
 	"github.com/andydunstall/piko/server"
 	"github.com/andydunstall/piko/server/cluster"
 	"github.com/andydunstall/piko/server/config"
@@ -15,25 +20,89 @@ import (
 type Node struct {
 	server *server.Server
 
+	rootCAPool *x509.CertPool
+
 	ctx    context.Context
 	cancel func()
 
 	wg sync.WaitGroup
 }
 
-func NewNode(join []string, logger log.Logger) *Node {
+func NewNode(opts ...Option) *Node {
+	options := options{
+		logger: log.NewNopLogger(),
+	}
+	for _, o := range opts {
+		o.apply(&options)
+	}
+
 	conf := config.Default()
 	conf.Cluster.NodeID = cluster.GenerateNodeID()
-	conf.Cluster.Join = join
+	conf.Cluster.Join = options.join
 	conf.Proxy.BindAddr = "127.0.0.1:0"
 	conf.Upstream.BindAddr = "127.0.0.1:0"
 	conf.Admin.BindAddr = "127.0.0.1:0"
 	conf.Gossip.BindAddr = "127.0.0.1:0"
 	conf.Gossip.Interval = time.Millisecond * 10
+	conf.Auth = options.authConfig
+
+	// If TLS is enabled, generate a certificate and root CA then write to a
+	// file.
+	var rootCAPool *x509.CertPool
+	if options.tls {
+		conf.Proxy.TLS.Enabled = true
+		conf.Upstream.TLS.Enabled = true
+		conf.Admin.TLS.Enabled = true
+
+		pool, cert, err := testutil.LocalTLSServerCert()
+		if err != nil {
+			panic("tls cert: " + err.Error())
+		}
+		rootCAPool = pool
+
+		f, err := os.CreateTemp("", "piko")
+		if err != nil {
+			panic("create temp: " + err.Error())
+		}
+
+		for _, certBytes := range cert.Certificate {
+			block := pem.Block{
+				Type:  "CERTIFICATE",
+				Bytes: certBytes,
+			}
+			if err := pem.Encode(f, &block); err != nil {
+				panic("encode pem: " + err.Error())
+			}
+		}
+
+		conf.Proxy.TLS.Cert = f.Name()
+		conf.Upstream.TLS.Cert = f.Name()
+		conf.Admin.TLS.Cert = f.Name()
+
+		f, err = os.CreateTemp("", "piko")
+		if err != nil {
+			panic("create temp: " + err.Error())
+		}
+
+		keyBytes := x509.MarshalPKCS1PrivateKey(cert.PrivateKey.(*rsa.PrivateKey))
+
+		block := pem.Block{
+			Type:  "RSA PRIVATE KEY",
+			Bytes: keyBytes,
+		}
+
+		if err := pem.Encode(f, &block); err != nil {
+			panic("encode pem: " + err.Error())
+		}
+
+		conf.Proxy.TLS.Key = f.Name()
+		conf.Upstream.TLS.Key = f.Name()
+		conf.Admin.TLS.Key = f.Name()
+	}
 
 	server, err := server.NewServer(
 		conf,
-		logger.With(zap.String("node", conf.Cluster.NodeID)),
+		options.logger.With(zap.String("node", conf.Cluster.NodeID)),
 	)
 	if err != nil {
 		panic("server: " + err.Error())
@@ -41,9 +110,10 @@ func NewNode(join []string, logger log.Logger) *Node {
 
 	ctx, cancel := context.WithCancel(context.Background())
 	return &Node{
-		server: server,
-		ctx:    ctx,
-		cancel: cancel,
+		server:     server,
+		rootCAPool: rootCAPool,
+		ctx:        ctx,
+		cancel:     cancel,
 	}
 }
 
@@ -65,6 +135,10 @@ func (n *Node) GossipAddr() string {
 
 func (n *Node) ClusterState() *cluster.State {
 	return n.server.ClusterState()
+}
+
+func (n *Node) RootCAPool() *x509.CertPool {
+	return n.rootCAPool
 }
 
 func (n *Node) Start() {
